+++ CHANGELOG.md
@@ -27,6 +27,16 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   a warning if the user does not specify target revision  explicitly. In the near
   future those commands will fail if target revision is not specified.
 
+* In the next release, `jj split` will no longer move bookmarks to the second
+  revision created by the split. Instead, local bookmarks associated with the
+  target revision will move to the first revision created by the split (which
+  inherits the target revision's change id). You can opt out of this change by
+  setting `split.legacy-bookmark-behavior = true`, but this will likely be
+  removed in a future release. You can also try the new behavior now by setting
+  `split.legacy-bookmark-behavior = false`.
+  [#3419](https://github.com/jj-vcs/jj/issues/3419)
+
+
 ### New features
 
 * `jj undo` now shows a hint when undoing an undo operation that the user may